requirements.txt
sentence-transformers
transformers
faiss-cpu
numpy
PyPDF2 